+++ nbc-nep-front/src/api/supabase/space.ts
@@ -1,6 +1,6 @@
 import { SpaceInfo } from "@/components/spaces/types/space.types";
 import { supabase } from "@/supabase/supabase";
-import { Tables, TablesInsert } from "@/types/supabase";
+import { Tables, TablesInsert } from "@/supabase/types/supabase";
 
 export const createSpaceHandler = async (
   spaceInfo: SpaceInfo
