+++ nbc-nep-front/src/components/metaverse/MetaverseComponent.tsx
@@ -2,13 +2,13 @@ import GlobalNavBar from "@/components/metaverse/globalNavBar/GlobalNavBar";
 import MetaverseChatBar from "@/components/metaverse/metaverseChat/metaverseChatBar/MetaverseChatBar";
 import MetaversePlayerList from "@/components/metaverse/metaversePlayerList/MetaversePlayerList";
 import { usePlayerContext } from "@/context/MetaversePlayerProvider";
-import { CharacterScenes } from "@/metaverse/scenes/characterScenes";
-import { ScenesMain } from "@/metaverse/scenes/scenesMain";
 import Phaser from "phaser";
 import { useEffect } from "react";
 import styled from "styled-components";
 import VideoConference from "../video-conference/VideoConference";
-import PhaserSceneManager from "@/metaverse/scenes/phaserSceneManager";
+import { SetupScene } from "@/components/metaverse/libs/setupScene";
+import { SceneClass } from "@/components/metaverse/libs/sceneClass";
+import PhaserSceneManager from "@/components/metaverse/libs/phaserSceneManager";
 
 const MetaverseComponent = () => {
   const { spaceId, playerSpaceInfoData, id, display_name } = usePlayerContext();
@@ -31,13 +31,13 @@ const MetaverseComponent = () => {
           default: "arcade",
           arcade: {
             gravity: { y: 0 },
-            debug: true,
+            debug: false,
             width: 1280,
             height: 800,
             fps: 60,
           },
         },
-        scene: [ScenesMain, CharacterScenes],
+        scene: [SetupScene, SceneClass],
       };
 
       game = new Phaser.Game(config);
