+++ nbc-nep-front/src/components/metaverse/libs/currentPlayer.ts
