+++ nbc-nep-front/src/components/metaverse/libs/phaserSceneManager.ts
