+++ nbc-nep-front/src/components/metaverse/metaverseChat/metaverseChatBar/MetaverseChatCard.tsx
