+++ nbc-nep-front/src/components/metaverse/metaverseChat/metaverseChatBar/MetaverseChatHeader.tsx
