+++ nbc-nep-front/src/components/metaverse/metaversePlayerList/MetaversePlayerCard.tsx
