+++ nbc-nep-front/src/components/spaces/CreateSpaceForm.tsx
@@ -7,7 +7,7 @@ import {
   UseFormRegister,
 } from "react-hook-form";
 import styled from "styled-components";
-import { FORM_CHARACTER } from "./constatns/constants";
+import { FORM_CHARACTER } from "@/components/spaces/constants/constants";
 import { Procedure, SpaceInfo } from "./types/space.types";
 
 /**
