+++ nbc-nep-front/src/components/spaces/constants/constants.ts
