+++ nbc-nep-front/src/components/video-conference/libs/dnd.ts
