+++ nbc-nep-front/src/components/video-conference/libs/util.ts
