+++ nbc-nep-front/src/components/video-conference/video-media-item/OtherPlayerShareMediaItem.tsx
@@ -4,7 +4,10 @@ import { layoutOpen } from "@/redux/modules/layoutSlice";
 import { Player } from "@/components/metaverse/types/metaverse";
 import { useRef } from "react";
 import ShareMediaItem from "../ShareMediaItem";
-import { isArrayEmpty, splitVideoSource } from "../lib/util";
+import {
+  isArrayEmpty,
+  splitVideoSource,
+} from "@/components/video-conference/libs/util";
 import {
   SPACING,
   StStackItem,
