+++ nbc-nep-front/src/components/video-conference/video-media-item/ShareMediaItemContainer.tsx
@@ -4,7 +4,10 @@ import { Player } from "@/components/metaverse/types/metaverse";
 import styled from "styled-components";
 import ShareMediaItem from "../ShareMediaItem";
 import ShareScreenContainer from "../ShareScreenContainer";
-import { isArrayEmpty, splitVideoSource } from "../lib/util";
+import {
+  isArrayEmpty,
+  splitVideoSource,
+} from "@/components/video-conference/libs/util";
 import { Producer } from "../types/ScreenShare.types";
 import DefaultShareMediaItem from "./DefaultShareMediaItem";
 import OtherPlayerShareMediaItem from "./OtherPlayerShareMediaItem";
