+++ nbc-nep-front/src/hooks/GNB/useGNBIconButtons.ts
@@ -2,7 +2,7 @@ import { useAppDispatch, useAppSelector } from "@/hooks/useReduxTK";
 import { IconButtonProperty } from "@/components/metaverse/globalNavBar/globalNavBarIconWrapper/iconButton/types/iconButtonTypes";
 import chartIcon from "@/assets/icons/Comments.svg";
 import { setIsSomeSection } from "@/redux/modules/globalNavBarSlice";
-import { ChatType } from "@/components/metaverse/metaverseChat/types/ChatType";
+import { ChatType } from "@/components/metaverse/types/ChatType";
 import { setIsOpenChat } from "@/redux/modules/chatTypeSlice";
 import SettingsIcon from "@/assets/icons/Cog.svg";
 import reportIcon from "@/assets/icons/User Headset.svg";
