+++ nbc-nep-front/src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import OnChangeUserSession from "@/components/layout/OnChangeUserSession";
-import StoreProvider from "@/supabase/StoreProvier";
+import StoreProvider from "@/redux/StoreProvier";
 import GlobalStyle, { theme } from "@/styles/Globalstyle";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { NextPage } from "next";
