+++ nbc-nep-front/src/supabase/types/supabase.tables.type.ts
